+++ PersonLibrary/Model/Teacher.cs
@@ -14,15 +14,10 @@ namespace PersonLibrary.Model
 
         public string Subject { get; set; }
 
-        public override string SPrintInfo()
+        public override string PrintInfo()
         {
-            return base.SPrintInfo();
+            return base.PrintInfo();
         }
-
-        /*public override void PrintInfo()
-        {
-            base.PrintInfo();
-        }*/
 
         public override string Header()
         {
