+++ StudentApp/PersonException.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace StudentApp
+{
+    internal class PersonException: Exception
+    {
+
+
+
+    }
+}
